+++ 1.CensorWords/CensorWords.py
@@ -1,0 +1,4 @@
+def censor_string(string, lst, char):
+ modified_string=string.split()
+ returned_words = [word if word not in lst else char*len(word) for word in modified_string]
+ print (' '.join(returned_words))
